requirements.txt
Django>=1.7,<1.8
steelscript>=0.9.5
steelscript.netprofiler>=0.9.4
djangorestframework==2.3.13
djangorestframework-csv==1.3.3
django-extensions==1.4.6
django-model-utils==2.0.3
jsonfield==0.9.20
numpy>=1.8.0,<2.0
pandas>=0.15.1,<0.16
pygeoip>=0.2.6
python-dateutil>=2.2
pytz>=2013.8
six>=1.3.0
wsgiref>=0.1.2
django-admin-tools==0.5.2
ansi2html>=1.0.6
django-ace==1.0.2
apscheduler>=3.0
celery>=3.1
django-celery==3.1.16
redis==2.10.3
flask==0.10.1
flask_restful==0.3.2
